requirements.txt
asgiref==3.5.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.12
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-registration==3.2
gunicorn==20.1.0
phonenumbers==8.12.44
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==6.0.0
